Update for Python 3
Python 3 differentiates strings and bytes.

I also move the .split() and eliminated what is essentially duplicate
logging of the incoming command.  If logging both the split and unsplit
version is desired, the previous code seems backwards anyway.  Rather
than logging it split and then re-joining it, it should be logged
unsplit and then after the split.